hw/sd/sdcard: Make iolen unsigned
I/O request length can not be negative.

Message-Id: <20200605102230.21493-9-philmd@redhat.com>
Supersedes: <20200604182502.24228-1-f4bug@amsat.org>

+++ hw/sd/sd.c
@@ -1950,7 +1950,7 @@ uint8_t sd_read_data(SDState *sd)
 {
     /* TODO: Append CRCs */
     uint8_t ret;
-    int io_len;
+    size_t io_len;
 
     if (!sd->blk || !blk_is_inserted(sd->blk) || !sd->enable)
         return 0x00;
